Add all project files
This is a project upload after the project was finished.

The project was done at Archives de l'Etat du Valais in Spring 2021. The goal was to create a pipeline to extract metadata from archive images from the collection Imprimés Valaisans.

Feel free to contact the author for any questions.